+++ src/Wrapper/Wrapper.vcxproj
@@ -157,9 +157,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="application.h" />
     <ClInclude Include="messages.h" />
-    <ClInclude Include="service.h" />
     <ClInclude Include="service_config.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
@@ -169,11 +167,9 @@
     <ClInclude Include="wrapper-help.h" />
     <ClInclude Include="wrapper-log.h" />
     <ClInclude Include="wrapper-memory.h" />
-    <ClInclude Include="wrapper.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="service.c" />
-    <ClCompile Include="service_config.c" />
     <ClCompile Include="service_control.c" />
     <ClCompile Include="stdafx.c">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
@@ -188,7 +184,6 @@
     <ClCompile Include="wrapper-help.c" />
     <ClCompile Include="wrapper-log.c" />
     <ClCompile Include="wrapper-memory.c" />
-    <ClCompile Include="wrapper.c" />
   </ItemGroup>
   <ItemGroup>
     <None Include="config.example" />
