--- src/Wrapper/application.h
@@ -1,3 +1,0 @@
-#pragma once
-
-
