--- src/Wrapper/service_config.c
@@ -1,484 +1,0 @@
-#include "stdafx.h"
-#include "messages.h"
-#include "service_config.h"
-#include "wrapper-log.h"
-#include "service.h"
-
-int wrapper_service_run(wrapper_config_t* config, wrapper_error_t** error)
-{
-	TCHAR module_path[_MAX_PATH];
-	GetModuleFileName(NULL, module_path, _MAX_PATH);
-	PathCchRenameExtension(module_path, sizeof(module_path) / sizeof(module_path[0]), _T(".log"));
-	wrapper_log_set_handler(wrapper_log_file_handler, module_path);
-
-	WRAPPER_INFO(_T("Starting Service"));
-	SERVICE_TABLE_ENTRY DispatchTable[] =
-	{
-		{ config->name, (LPSERVICE_MAIN_FUNCTION)SvcMain },
-		{ NULL, NULL }
-	};
-
-	if (!StartServiceCtrlDispatcher(DispatchTable))
-	{
-		WRAPPER_ERROR(_T("Failed to start service"));
-		SvcReportEvent(TEXT("StartServiceCtrlDispatcher"));
-	}
-	WRAPPER_INFO(_T("Done"));
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Installs a service in the SCM database
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_install(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager;
-	SC_HANDLE schService;
-	TCHAR szPath[MAX_PATH];
-
-	if (!GetModuleFileName(NULL, szPath, MAX_PATH))
-	{
-		printf("Cannot install service (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Create the service
-
-	schService = CreateService(
-		schSCManager,              // SCM database 
-		config->name,                   // name of service 
-		config->title,                   // service name to display 
-		SERVICE_ALL_ACCESS,        // desired access 
-		SERVICE_WIN32_OWN_PROCESS, // service type 
-		SERVICE_DEMAND_START,      // start type 
-		SERVICE_ERROR_NORMAL,      // error control type 
-		szPath,                    // path to service's binary 
-		NULL,                      // no load ordering group 
-		NULL,                      // no tag identifier 
-		NULL,                      // no dependencies 
-		NULL,                      // LocalSystem account 
-		NULL);                     // no password 
-
-	if (schService == NULL)
-	{
-		printf("CreateService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-	else printf("Service installed successfully\n");
-
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Retrieves and displays the current service configuration.
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_query(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager = NULL;
-	SC_HANDLE schService = NULL;
-	LPQUERY_SERVICE_CONFIG lpsc = NULL;
-	LPSERVICE_DESCRIPTION lpsd = NULL;
-	DWORD dwBytesNeeded = 0, cbBufSize = 0, dwError = 0;
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the service.
-
-	schService = OpenService(
-		schSCManager,          // SCM database 
-		config->name,             // name of service 
-		SERVICE_QUERY_CONFIG); // need query config access 
-
-	if (schService == NULL)
-	{
-		printf("OpenService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-
-	// Get the configuration information.
-
-	if (!QueryServiceConfig(
-		schService,
-		NULL,
-		0,
-		&dwBytesNeeded))
-	{
-		dwError = GetLastError();
-		if (ERROR_INSUFFICIENT_BUFFER == dwError)
-		{
-			cbBufSize = dwBytesNeeded;
-			lpsc = (LPQUERY_SERVICE_CONFIG)LocalAlloc(LMEM_FIXED, cbBufSize);
-		}
-		else
-		{
-			printf("QueryServiceConfig failed (%d)", dwError);
-			goto cleanup;
-		}
-	}
-
-	if (!QueryServiceConfig(
-		schService,
-		lpsc,
-		cbBufSize,
-		&dwBytesNeeded))
-	{
-		printf("QueryServiceConfig failed (%d)", GetLastError());
-		goto cleanup;
-	}
-
-	if (!QueryServiceConfig2(
-		schService,
-		SERVICE_CONFIG_DESCRIPTION,
-		NULL,
-		0,
-		&dwBytesNeeded))
-	{
-		dwError = GetLastError();
-		if (ERROR_INSUFFICIENT_BUFFER == dwError)
-		{
-			cbBufSize = dwBytesNeeded;
-			lpsd = (LPSERVICE_DESCRIPTION)LocalAlloc(LMEM_FIXED, cbBufSize);
-		}
-		else
-		{
-			printf("QueryServiceConfig2 failed (%d)", dwError);
-			goto cleanup;
-		}
-	}
-
-	if (!QueryServiceConfig2(
-		schService,
-		SERVICE_CONFIG_DESCRIPTION,
-		(LPBYTE)lpsd,
-		cbBufSize,
-		&dwBytesNeeded))
-	{
-		printf("QueryServiceConfig2 failed (%d)", GetLastError());
-		goto cleanup;
-	}
-
-	// Print the configuration information.
-
-	_tprintf(TEXT("%s configuration: \n"), config->name);
-	_tprintf(TEXT("  Type: 0x%x\n"), lpsc->dwServiceType);
-	_tprintf(TEXT("  Start Type: 0x%x\n"), lpsc->dwStartType);
-	_tprintf(TEXT("  Error Control: 0x%x\n"), lpsc->dwErrorControl);
-	_tprintf(TEXT("  Binary path: %s\n"), lpsc->lpBinaryPathName);
-	_tprintf(TEXT("  Account: %s\n"), lpsc->lpServiceStartName);
-
-	if (lpsd->lpDescription != NULL && lstrcmp(lpsd->lpDescription, TEXT("")) != 0)
-		_tprintf(TEXT("  Description: %s\n"), lpsd->lpDescription);
-	if (lpsc->lpLoadOrderGroup != NULL && lstrcmp(lpsc->lpLoadOrderGroup, TEXT("")) != 0)
-		_tprintf(TEXT("  Load order group: %s\n"), lpsc->lpLoadOrderGroup);
-	if (lpsc->dwTagId != 0)
-		_tprintf(TEXT("  Tag ID: %d\n"), lpsc->dwTagId);
-	if (lpsc->lpDependencies != NULL && lstrcmp(lpsc->lpDependencies, TEXT("")) != 0)
-		_tprintf(TEXT("  Dependencies: %s\n"), lpsc->lpDependencies);
-
-	LocalFree(lpsc);
-	LocalFree(lpsd);
-
-cleanup:
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Disables the service.
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_disable(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager;
-	SC_HANDLE schService;
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the service.
-
-	schService = OpenService(
-		schSCManager,            // SCM database 
-		config->name,               // name of service 
-		SERVICE_CHANGE_CONFIG);  // need change config access 
-
-	if (schService == NULL)
-	{
-		printf("OpenService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-
-	// Change the service start type.
-
-	if (!ChangeServiceConfig(
-		schService,        // handle of service 
-		SERVICE_NO_CHANGE, // service type: no change 
-		SERVICE_DISABLED,  // service start type 
-		SERVICE_NO_CHANGE, // error control: no change 
-		NULL,              // binary path: no change 
-		NULL,              // load order group: no change 
-		NULL,              // tag ID: no change 
-		NULL,              // dependencies: no change 
-		NULL,              // account name: no change 
-		NULL,              // password: no change 
-		NULL))            // display name: no change
-	{
-		printf("ChangeServiceConfig failed (%d)\n", GetLastError());
-	}
-	else printf("Service disabled successfully.\n");
-
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Enables the service.
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_enable(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager;
-	SC_HANDLE schService;
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the service.
-
-	schService = OpenService(
-		schSCManager,            // SCM database 
-		config->name,               // name of service 
-		SERVICE_CHANGE_CONFIG);  // need change config access 
-
-	if (schService == NULL)
-	{
-		printf("OpenService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-
-	// Change the service start type.
-
-	if (!ChangeServiceConfig(
-		schService,            // handle of service 
-		SERVICE_NO_CHANGE,     // service type: no change 
-		SERVICE_DEMAND_START,  // service start type 
-		SERVICE_NO_CHANGE,     // error control: no change 
-		NULL,                  // binary path: no change 
-		NULL,                  // load order group: no change 
-		NULL,                  // tag ID: no change 
-		NULL,                  // dependencies: no change 
-		NULL,                  // account name: no change 
-		NULL,                  // password: no change 
-		NULL))                // display name: no change
-	{
-		printf("ChangeServiceConfig failed (%d)\n", GetLastError());
-	}
-	else printf("Service enabled successfully.\n");
-
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Updates the service description to "This is a test description".
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_update(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager;
-	SC_HANDLE schService;
-	SERVICE_DESCRIPTION sd;
-	LPTSTR szDesc = config->description;
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the service.
-
-	schService = OpenService(
-		schSCManager,            // SCM database 
-		config->name,               // name of service 
-		SERVICE_CHANGE_CONFIG);  // need change config access 
-
-	if (schService == NULL)
-	{
-		printf("OpenService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-
-	// Change the service description.
-
-	sd.lpDescription = szDesc;
-
-	if (!ChangeServiceConfig2(
-		schService,                 // handle to service
-		SERVICE_CONFIG_DESCRIPTION, // change: description
-		&sd))                      // new description
-	{
-		printf("ChangeServiceConfig2 failed\n");
-	}
-	else printf("Service description updated successfully.\n");
-
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
-
-//
-// Purpose: 
-//   Deletes a service from the SCM database
-//
-// Parameters:
-//   None
-// 
-// Return value:
-//   None
-//
-int wrapper_service_delete(wrapper_config_t* config, wrapper_error_t** error)
-{
-	SC_HANDLE schSCManager;
-	SC_HANDLE schService;
-	SERVICE_STATUS ssStatus;
-
-	// Get a handle to the SCM database. 
-
-	schSCManager = OpenSCManager(
-		NULL,                    // local computer
-		NULL,                    // ServicesActive database 
-		SC_MANAGER_ALL_ACCESS);  // full access rights 
-
-	if (NULL == schSCManager)
-	{
-		printf("OpenSCManager failed (%d)\n", GetLastError());
-		return;
-	}
-
-	// Get a handle to the service.
-
-	schService = OpenService(
-		schSCManager,       // SCM database 
-		config->name,          // name of service 
-		DELETE);            // need delete access 
-
-	if (schService == NULL)
-	{
-		printf("OpenService failed (%d)\n", GetLastError());
-		CloseServiceHandle(schSCManager);
-		return;
-	}
-
-	// Delete the service.
-
-	if (!DeleteService(schService))
-	{
-		printf("DeleteService failed (%d)\n", GetLastError());
-	}
-	else printf("Service deleted successfully\n");
-
-	CloseServiceHandle(schService);
-	CloseServiceHandle(schSCManager);
-
-	return 1;
-}
