--- src/Wrapper/wrapper.c
@@ -1,8 +1,0 @@
-#include "stdafx.h"
-#include "messages.h"
-
-HRESULT GetServiceName(_Out_ LPTSTR  pszDest, _In_  size_t  cbDest)
-{
-	//TODO: Read from Configuration File
-	return StringCbCopy(pszDest, cbDest, _T("MyWrapperService"));
-}
