--- src/Wrapper/wrapper.h
@@ -1,3 +1,0 @@
-#pragma once
-
-BOOL GetServiceName(_Out_ LPTSTR  pszDest, _In_  size_t  cbDest);
